Change project structure (add main directory "com.bring.project")
