--- src/main/java/annotation/Configuration.java
@@ -1,4 +1,0 @@
-package annotation;
-
-public @interface Configuration {
-}
