+++ src/main/java/com/bring/project/App.java
@@ -1,3 +1,5 @@
+package com.bring.project;
+
 public class App {
     public static void main(String[] args) {
