+++ src/main/java/com/bring/project/annotation/Configuration.java
@@ -1,0 +1,4 @@
+package com.bring.project.annotation;
+
+public @interface Configuration {
+}
