+++ src/main/java/com/bring/project/configuration/AppConfiguration.java
@@ -1,0 +1,4 @@
+package com.bring.project.configuration;
+
+public class AppConfiguration {
+}
