+++ src/main/java/com/bring/project/exception/BoboNotFoundException.java
@@ -1,4 +1,4 @@
-package exception;
+package com.bring.project.exception;
 
 public class BoboNotFoundException {
 }
