+++ src/main/java/com/bring/project/service/creator/BoboCreatorService.java
@@ -1,0 +1,4 @@
+package com.bring.project.service.creator;
+
+public interface BoboCreatorService {
+}
