+++ src/main/java/com/bring/project/service/creator/impl/BoboCreatorServiceImpl.java
@@ -1,0 +1,6 @@
+package com.bring.project.service.creator.impl;
+
+import com.bring.project.service.creator.BoboCreatorService;
+
+public class BoboCreatorServiceImpl implements BoboCreatorService {
+}
