+++ src/main/java/com/bring/project/service/processor/AnnotationProcessorService.java
@@ -1,4 +1,4 @@
-package service.processor;
+package com.bring.project.service.processor;
 
 public interface AnnotationProcessorService {
 }
