+++ src/main/java/com/bring/project/service/processor/impl/AnnotationProcessorServiceImpl.java
@@ -1,0 +1,6 @@
+package com.bring.project.service.processor.impl;
+
+import com.bring.project.service.processor.AnnotationProcessorService;
+
+public class AnnotationProcessorServiceImpl implements AnnotationProcessorService {
+}
