--- src/main/java/configuration/AppConfiguration.java
@@ -1,4 +1,0 @@
-package configuration;
-
-public class AppConfiguration {
-}
