--- src/main/java/service/creator/BoboCreatorService.java
@@ -1,4 +1,0 @@
-package service.creator;
-
-public interface BoboCreatorService {
-}
