--- src/main/java/service/creator/impl/BoboCreatorServiceImpl.java
@@ -1,6 +1,0 @@
-package service.creator.impl;
-
-import service.creator.BoboCreatorService;
-
-public class BoboCreatorServiceImpl implements BoboCreatorService {
-}
