--- src/main/java/service/processor/impl/AnnotationProcessorServiceImpl.java
@@ -1,6 +1,0 @@
-package service.processor.impl;
-
-import service.processor.AnnotationProcessorService;
-
-public class AnnotationProcessorServiceImpl implements AnnotationProcessorService {
-}
